cmake: remove CMAKE_INSTALL_RPATH
The CMake builder in Spack actually adds incorrect rpaths. They are
unfiltered and incorrectly ordered compared to what the compiler wrapper
adds.

There is no need to set CMAKE_INSTALL_RPATH because CMake preserves
"toolchain" rpaths, which are rpaths injected automatically, and the
compiler wrapper is from CMake's perspective part of the toolchain.